refactor(featureflag): move FeatureFlagsScreen to impl, delegate via BlocScreen
Same recipe as the other leaves. featureflag/public gains
publicResClass = true; featureflag/impl gains the Compose plugin
+ ui/text deps. RootBloc.Child.FeatureFlags delegates via
BlocScreen by bloc.
