+++ client/featureflag/impl/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     alias(libs.plugins.kmpLibrary)
+    alias(libs.plugins.compose)
     alias(libs.plugins.kotlinSerialization)
 }
 
@@ -7,12 +8,16 @@ kotlin {
     sourceSets {
         commonMain.dependencies {
             implementation(projects.client.featureflag.public)
+            implementation(projects.client.text.public)
+            implementation(projects.client.ui.public)
             implementation(projects.client.shared)
             implementation(projects.client.auth.data.public)
             implementation(libs.arkivanov.decompose.core)
+            implementation(libs.arkivanov.decompose.compose.extensions)
             implementation(libs.kotlinx.serialization.json)
             implementation(libs.supabase.client)
             implementation(libs.supabase.postgrest)
+            implementation(compose.components.resources)
         }
         commonTest.dependencies {
             implementation(libs.multiplatform.settings.test)
