+++ client/featureflag/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/featureflag/impl/FeatureFlagsBlocImpl.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.featureflag.impl
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.di.AppScope
@@ -9,6 +11,7 @@ import com.plusmobileapps.chefmate.featureflag.FeatureFlagsBloc
 import com.plusmobileapps.chefmate.featureflag.FeatureFlagsBloc.Output
 import com.plusmobileapps.chefmate.featureflag.Override
 import com.plusmobileapps.chefmate.featureflag.StringFlag
+import com.plusmobileapps.chefmate.featureflag.impl.ui.FeatureFlagsScreen
 import com.plusmobileapps.chefmate.getViewModel
 import com.plusmobileapps.metro.extensions.assistedfactory.ContributesAssistedFactory
 import dev.zacsweers.metro.Assisted
@@ -49,5 +52,10 @@ class FeatureFlagsBlocImpl(
 
     override fun onClearAllOverrides() {
         viewModel.clearAll()
+    }
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        FeatureFlagsScreen(bloc = this, modifier = modifier)
     }
 }
